backend/requirements.txt
# Agent orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0

# LLM provider
groq>=0.11.0

# Embeddings + vector store (local, no API key required)
sentence-transformers>=3.0.0
chromadb>=0.5.0

# API layer
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
strawberry-graphql[fastapi]>=0.243.0

# Quick-test UI
streamlit>=1.40.0

# Data validation + state
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Observability
langfuse>=4.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
